shotgun works, making sure player movement is clamped

+++ Assets/Scripts/Player/Combat/PlayerProjectileScript.cs
@@ -11,7 +11,7 @@ public class PlayerProjectileScript : MonoBehaviour
     [SerializeField] GameObject breakParticle, muzzleEffect, normalHitFX, critHitFX, homingFX, teleportFX; // the particle we use on death
 
     RaycastHit hit; // our raycast hit
-    [SerializeField] int deathTime = 30;
+    [SerializeField] float deathTime = 30;
 
     [SerializeField] bool usesPhysics;
     [SerializeField] float physicsLaunchForce;
